// types.ts
export type FooterItemT = {
  svg_source: string;
  svg_alt: string;
  title: string;
  text: string;
  link: string;
  header: string;
};

export type FooterLogoT = {
  svg_src: string;
  svg_alt: string;
  link: string;
  ratio: string;
  mobile_width: number;
  desktop_width: number;
};

export type ClientImageT = {
  image_src: string;
  image_alt: string;
};